requirements.txt
spotipy>=2.25.0
ytmusicapi>=1.10.0,<1.11.0
python-dotenv>=1.0.0
